// lib/types.d.ts
export interface paths {
// ...
  "/systems/{systemSymbol}/waypoints": {
    /**
     * List Waypoints in System
     * @description Return a paginated list of all of the waypoints for a given system.
     *
     * If a waypoint is uncharted, it will return the `Uncharted` trait instead of its actual traits.
     */
    get: operations["get-system-waypoints"];
    parameters: {
      path: {
        /** @description The system symbol */
        systemSymbol: string;
      };
    };
  };
// ...
  "/systems/{systemSymbol}/waypoints/{waypointSymbol}/market": {
    /**
     * Get Market
     * @description Retrieve imports, exports and exchange data from a marketplace. Requires a waypoint that has the `Marketplace` trait to use.
     *
     * Send a ship to the waypoint to access trade good prices and recent transactions. Refer to the [Market Overview page](https://docs.spacetraders.io/game-concepts/markets) to gain better a understanding of the market in the game.
     */
    get: operations["get-market"];
    parameters: {
      path: {
        /** @description The system symbol */
        systemSymbol: string;
        /** @description The waypoint symbol */
        waypointSymbol: string;
      };
    };
  };
  "/systems/{systemSymbol}/waypoints/{waypointSymbol}/shipyard": {
    /**
     * Get Shipyard
     * @description Get the shipyard for a waypoint. Requires a waypoint that has the `Shipyard` trait to use. Send a ship to the waypoint to access data on ships that are currently available for purchase and recent transactions.
     */
    get: operations["get-shipyard"];
    parameters: {
      path: {
        /** @description The system symbol */
        systemSymbol: string;
        /** @description The waypoint symbol */
        waypointSymbol: string;
      };
    };
  };
// ...
  "/my/agent": {
    /**
     * Get Agent
     * @description Fetch your agent's details.
     */
    get: operations["get-my-agent"];
  };
  "/agents": {
    /**
     * List Agents
     * @description Fetch agents details.
     */
    get: operations["get-agents"];
  };
  "/agents/{agentSymbol}": {
    /**
     * Get Public Agent
     * @description Fetch agent details.
     */
    get: operations["get-agent"];
  };
  "/my/contracts": {
    /**
     * List Contracts
     * @description Return a paginated list of all your contracts.
     */
    get: operations["get-contracts"];
  };
  "/my/contracts/{contractId}": {
    /**
     * Get Contract
     * @description Get the details of a contract by ID.
     */
    get: operations["get-contract"];
    parameters: {
      path: {
        /** @description The contract ID */
        contractId: string;
      };
    };
  };
  "/my/contracts/{contractId}/accept": {
    /**
     * Accept Contract
     * @description Accept a contract by ID.
     *
     * You can only accept contracts that were offered to you, were not accepted yet, and whose deadlines has not passed yet.
     */
    post: operations["accept-contract"];
    parameters: {
      path: {
        /** @description The contract ID to accept. */
        contractId: string;
      };
    };
  };
// ...
  "/my/ships": {
    /**
     * List Ships
     * @description Return a paginated list of all of ships under your agent's ownership.
     */
    get: operations["get-my-ships"];
    /**
     * Purchase Ship
     * @description Purchase a ship from a Shipyard. In order to use this function, a ship under your agent's ownership must be in a waypoint that has the `Shipyard` trait, and the Shipyard must sell the type of the desired ship.
     *
     * Shipyards typically offer ship types, which are predefined templates of ships that have dedicated roles. A template comes with a preset of an engine, a reactor, and a frame. It may also include a few modules and mounts.
     */
    post: operations["purchase-ship"];
  };
// ...
  "/my/ships/{shipSymbol}/orbit": {
    /**
     * Orbit Ship
     * @description Attempt to move your ship into orbit at its current location. The request will only succeed if your ship is capable of moving into orbit at the time of the request.
     *
     * Orbiting ships are able to do actions that require the ship to be above surface such as navigating or extracting, but cannot access elements in their current waypoint, such as the market or a shipyard.
     *
     * The endpoint is idempotent - successive calls will succeed even if the ship is already in orbit.
     */
    post: operations["orbit-ship"];
    parameters: {
      path: {
        /** @description The symbol of the ship. */
        shipSymbol: string;
      };
    };
  };
// ...
  "/my/ships/{shipSymbol}/survey": {
    /**
     * Create Survey
     * @description Create surveys on a waypoint that can be extracted such as asteroid fields. A survey focuses on specific types of deposits from the extracted location. When ships extract using this survey, they are guaranteed to procure a high amount of one of the goods in the survey.
     *
     * In order to use a survey, send the entire survey details in the body of the extract request.
     *
     * Each survey may have multiple deposits, and if a symbol shows up more than once, that indicates a higher chance of extracting that resource.
     *
     * Your ship will enter a cooldown after surveying in which it is unable to perform certain actions. Surveys will eventually expire after a period of time or will be exhausted after being extracted several times based on the survey's size. Multiple ships can use the same survey for extraction.
     *
     * A ship must have the `Surveyor` mount installed in order to use this function.
     */
    post: operations["create-survey"];
    parameters: {
      path: {
        /** @description The symbol of the ship. */
        shipSymbol: string;
      };
    };
  };
  "/my/ships/{shipSymbol}/extract": {
    /**
     * Extract Resources
     * @description Extract resources from a waypoint that can be extracted, such as asteroid fields, into your ship. Send an optional survey as the payload to target specific yields.
     *
     * The ship must be in orbit to be able to extract and must have mining equipments installed that can extract goods, such as the `Gas Siphon` mount for gas-based goods or `Mining Laser` mount for ore-based goods.
     *
     * The survey property is now deprecated. See the `extract/survey` endpoint for more details.
     */
    post: operations["extract-resources"];
    parameters: {
      path: {
        /** @description The ship symbol. */
        shipSymbol: string;
      };
    };
  };
// ...
  "/my/ships/{shipSymbol}/navigate": {
    /**
     * Navigate Ship
     * @description Navigate to a target destination. The ship must be in orbit to use this function. The destination waypoint must be within the same system as the ship's current location. Navigating will consume the necessary fuel from the ship's manifest based on the distance to the target waypoint.
     *
     * The returned response will detail the route information including the expected time of arrival. Most ship actions are unavailable until the ship has arrived at it's destination.
     *
     * To travel between systems, see the ship's Warp or Jump actions.
     */
    post: operations["navigate-ship"];
    parameters: {
      path: {
        /** @description The ship symbol. */
        shipSymbol: string;
      };
    };
  };
// ...
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: {
// ...
    ShipType: "SHIP_PROBE" | "SHIP_MINING_DRONE" | "SHIP_SIPHON_DRONE" | "SHIP_INTERCEPTOR" | "SHIP_LIGHT_HAULER" | "SHIP_COMMAND_FRIGATE" | "SHIP_EXPLORER" | "SHIP_HEAVY_FREIGHTER" | "SHIP_LIGHT_SHUTTLE" | "SHIP_ORE_HOUND" | "SHIP_REFINING_FREIGHTER" | "SHIP_SURVEYOR";
// ...
    Waypoint: {
      symbol: components["schemas"]["WaypointSymbol"];
      type: components["schemas"]["WaypointType"];
      systemSymbol: components["schemas"]["SystemSymbol"];
      /** @description Relative position of the waypoint on the system's x axis. This is not an absolute position in the universe. */
      x: number;
      /** @description Relative position of the waypoint on the system's y axis. This is not an absolute position in the universe. */
      y: number;
      /** @description Waypoints that orbit this waypoint. */
      orbitals: components["schemas"]["WaypointOrbital"][];
      /** @description The symbol of the parent waypoint, if this waypoint is in orbit around another waypoint. Otherwise this value is undefined. */
      orbits?: string;
      faction?: components["schemas"]["WaypointFaction"];
      /** @description The traits of the waypoint. */
      traits: components["schemas"]["WaypointTrait"][];
      /** @description The modifiers of the waypoint. */
      modifiers?: components["schemas"]["WaypointModifier"][];
      chart?: components["schemas"]["Chart"];
      /** @description True if the waypoint is under construction. */
      isUnderConstruction: boolean;
    };
// ...
    WaypointTraitSymbol: "UNCHARTED" | "UNDER_CONSTRUCTION" | "MARKETPLACE" | "SHIPYARD" | "OUTPOST" | "SCATTERED_SETTLEMENTS" | "SPRAWLING_CITIES" | "MEGA_STRUCTURES" | "PIRATE_BASE" | "OVERCROWDED" | "HIGH_TECH" | "CORRUPT" | "BUREAUCRATIC" | "TRADING_HUB" | "INDUSTRIAL" | "BLACK_MARKET" | "RESEARCH_FACILITY" | "MILITARY_BASE" | "SURVEILLANCE_OUTPOST" | "EXPLORATION_OUTPOST" | "MINERAL_DEPOSITS" | "COMMON_METAL_DEPOSITS" | "PRECIOUS_METAL_DEPOSITS" | "RARE_METAL_DEPOSITS" | "METHANE_POOLS" | "ICE_CRYSTALS" | "EXPLOSIVE_GASES" | "STRONG_MAGNETOSPHERE" | "VIBRANT_AURORAS" | "SALT_FLATS" | "CANYONS" | "PERPETUAL_DAYLIGHT" | "PERPETUAL_OVERCAST" | "DRY_SEABEDS" | "MAGMA_SEAS" | "SUPERVOLCANOES" | "ASH_CLOUDS" | "VAST_RUINS" | "MUTATED_FLORA" | "TERRAFORMED" | "EXTREME_TEMPERATURES" | "EXTREME_PRESSURE" | "DIVERSE_LIFE" | "SCARCE_LIFE" | "FOSSILS" | "WEAK_GRAVITY" | "STRONG_GRAVITY" | "CRUSHING_GRAVITY" | "TOXIC_ATMOSPHERE" | "CORROSIVE_ATMOSPHERE" | "BREATHABLE_ATMOSPHERE" | "THIN_ATMOSPHERE" | "JOVIAN" | "ROCKY" | "VOLCANIC" | "FROZEN" | "SWAMP" | "BARREN" | "TEMPERATE" | "JUNGLE" | "OCEAN" | "RADIOACTIVE" | "MICRO_GRAVITY_ANOMALIES" | "DEBRIS_CLUSTER" | "DEEP_CRATERS" | "SHALLOW_CRATERS" | "UNSTABLE_COMPOSITION" | "HOLLOWED_INTERIOR" | "STRIPPED";
    /**
     * @description The type of waypoint.
     * @enum {string}
     */
    WaypointType: "PLANET" | "GAS_GIANT" | "MOON" | "ORBITAL_STATION" | "JUMP_GATE" | "ASTEROID_FIELD" | "ASTEROID" | "ENGINEERED_ASTEROID" | "ASTEROID_BASE" | "NEBULA" | "DEBRIS_FIELD" | "GRAVITY_WELL" | "ARTIFICIAL_GRAVITY_WELL" | "FUEL_STATION";
  };
  responses: never;
  parameters: never;
  requestBodies: {
  };
  headers: never;
  pathItems: never;
}